Se resuelven ejercicios de Tarea

+++ src/gNumericTypes/Overflow.java
@@ -4,8 +4,8 @@ public class Overflow {
 
 	public static void main(String args[])
 	{
-		int n = 1;
-	    for (int i = 0; i < 64; i++)
+		long n = 1;
+	    for (long i = 0; i < 64; i++)
 	    {
 	    	System.out.print("n: " + n + "\n");
 	        n = n * 2;
